+++ ide/tools/idetools/test/test_AbjadIDE_interpret_preface.py
@@ -1,8 +1,6 @@
-import abjad
 import ide
 import pathlib
-abjad_ide = ide.tools.idetools.AbjadIDE(is_test=True)
-configuration = ide.tools.idetools.AbjadIDEConfiguration()
+abjad_ide = ide.AbjadIDE(is_test=True)
 
 
 def test_AbjadIDE_interpret_preface_01():
@@ -10,29 +8,29 @@ def test_AbjadIDE_interpret_preface_01():
     '''
 
     tex_path = pathlib.Path(
-        configuration.abjad_ide_example_scores_directory,
-        'red_example_score',
-        'red_example_score',
+        abjad_ide.configuration.example_scores_directory,
+        'red_score',
+        'red_score',
         'build',
-        'letter-portrait',
+        'letter',
         'preface.tex',
         )
     pdf_path = pathlib.Path(
-        configuration.abjad_ide_example_scores_directory,
-        'red_example_score',
-        'red_example_score',
+        abjad_ide.configuration.example_scores_directory,
+        'red_score',
+        'red_score',
         'build',
-        'letter-portrait',
+        'letter',
         'preface.pdf',
         )
 
-    with abjad.FilesystemState(keep=[tex_path, pdf_path]):
-        pdf_path.unlink()
+    with ide.Test():
+        if pdf_path.exists():
+            pdf_path.unlink()
         assert not pdf_path.exists()
-        input_ = 'red~example~score bb letter-portrait pi q'
+        input_ = 'red~score bb letter pi q'
         abjad_ide._start(input_=input_)
         assert pdf_path.is_file()
-        assert abjad.TestManager._compare_backup(pdf_path)
 
 
 def test_AbjadIDE_interpret_preface_02():
@@ -40,24 +38,24 @@ def test_AbjadIDE_interpret_preface_02():
     '''
 
     tex_path = pathlib.Path(
-        configuration.abjad_ide_example_scores_directory,
-        'red_example_score',
-        'red_example_score',
+        abjad_ide.configuration.example_scores_directory,
+        'red_score',
+        'red_score',
         'build',
-        'letter-portrait',
+        'letter',
         'preface.tex',
         )
     pdf_path = pathlib.Path(
-        configuration.abjad_ide_example_scores_directory,
-        'red_example_score',
-        'red_example_score',
+        abjad_ide.configuration.example_scores_directory,
+        'red_score',
+        'red_score',
         'build',
-        'letter-portrait',
+        'letter',
         'preface.pdf',
         )
 
-    with abjad.FilesystemState(keep=[tex_path, pdf_path]):
+    with ide.Test():
         assert pdf_path.exists()
-        input_ = 'red~example~score bb letter-portrait pi q'
+        input_ = 'red~score bb letter pi q'
         abjad_ide._start(input_=input_)
         assert pdf_path.exists()
